refactor: applicant 관련 handler와 repository를 하나의 도메인 단위로 통합
